+++ ocaml_version.ml
@@ -125,6 +125,8 @@ module Releases = struct
   let all = [ v4_00; v4_01; v4_02; v4_03; v4_04;
               v4_05; v4_06; v4_07; v4_08; v4_09; v4_10 ]
 
+  let unreleased_betas = [ of_string_exn "4.10.0+beta1" ]
+
   let recent = [ v4_02; v4_03; v4_04; v4_05; v4_06; v4_07; v4_08; v4_09 ]
 
   let latest = v4_09